add permissions: owner, admin, member

+++ backend/internal/storage/company.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"context"
 
+	"github.com/gmohmad/diploma/internal/config"
 	"github.com/gmohmad/diploma/internal/models/domain"
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
@@ -36,12 +37,15 @@ func (s *Storage) CreateCompany(ctx context.Context, userID uuid.UUID, name stri
 	return company, nil
 }
 
-func (s *Storage) UpdateCompany(ctx context.Context, id uuid.UUID, name string) (*domain.Company, error) {
+func (s *Storage) UpdateCompany(ctx context.Context, userID, companyID uuid.UUID, name string) (*domain.Company, error) {
+	if !s.CheckPermission(ctx, userID, companyID, config.OwnerRole) {
+		return nil, domain.ErrInsufficientPermissions
+	}
 	query := `UPDATE companies
 	          SET name = $2, updated_at = NOW()
 	          WHERE id = $1
 	          RETURNING id, name, created_at, updated_at`
-	rows, err := s.client.Query(ctx, query, name)
+	rows, err := s.client.Query(ctx, query, companyID, name)
 	if err != nil {
 		return nil, err
 	}
@@ -49,9 +53,12 @@ func (s *Storage) UpdateCompany(ctx context.Context, id uuid.UUID, name string) 
 	return pgx.CollectOneRow(rows, pgx.RowToAddrOfStructByPos[domain.Company])
 }
 
-func (s *Storage) DeleteCompany(ctx context.Context, id uuid.UUID) error {
+func (s *Storage) DeleteCompany(ctx context.Context, userID, companyID uuid.UUID) error {
+	if !s.CheckPermission(ctx, userID, companyID, config.OwnerRole) {
+		return domain.ErrInsufficientPermissions
+	}
 	query := `DELETE FROM companies WHERE id = $1`
-	commandTag, err := s.client.Exec(ctx, query, id)
+	commandTag, err := s.client.Exec(ctx, query, companyID)
 	if err != nil {
 		return err
 	}
@@ -61,9 +68,12 @@ func (s *Storage) DeleteCompany(ctx context.Context, id uuid.UUID) error {
 	return nil
 }
 
-func (s *Storage) GetCompanyByID(ctx context.Context, id uuid.UUID) (*domain.Company, error) {
+func (s *Storage) GetCompanyByID(ctx context.Context, userID, companyID uuid.UUID) (*domain.Company, error) {
+	if !s.CheckPermission(ctx, userID, companyID, config.OwnerRole) {
+		return nil, domain.ErrInsufficientPermissions
+	}
 	query := `SELECT id, name, created_at, updated_at FROM companies WHERE id = $1`
-	rows, err := s.client.Query(ctx, query, id)
+	rows, err := s.client.Query(ctx, query, companyID)
 	if err != nil {
 		return nil, err
 	}
